Make gallium compile in win32.
Use FREE, MALLOC, CALLOC, GETENV wrappers.
Silence compiler warnings.
Add proper copyrights.

+++ src/mesa/pipe/draw/draw_vertex_fetch.c
@@ -64,16 +64,16 @@ fetch_attrib4(const void *ptr, unsigned format, float attrib[4])
       break;
 
    case PIPE_FORMAT_R32G32B32A32_SSCALED:
-      attrib[3] = ((int *) ptr)[3];
+      attrib[3] = (float) ((int *) ptr)[3];
       /* fall-through */
    case PIPE_FORMAT_R32G32B32_SSCALED:
-      attrib[2] = ((int *) ptr)[2];
+      attrib[2] = (float) ((int *) ptr)[2];
       /* fall-through */
    case PIPE_FORMAT_R32G32_SSCALED:
-      attrib[1] = ((int *) ptr)[1];
+      attrib[1] = (float) ((int *) ptr)[1];
       /* fall-through */
    case PIPE_FORMAT_R32_SSCALED:
-      attrib[0] = ((int *) ptr)[0];
+      attrib[0] = (float) ((int *) ptr)[0];
       break;
 
    default:
